requirements.txt
Django
django-bootstrap-breadcrumbs
django-debug-toolbar
django-extensions
django-js-reverse
djangorestframework
markdown
django-filter
django-environ
django-extensions
Werkzeug
#Pillow
psycopg2
django-redis
redis
requests
gevent
gunicorn
django-anymail
raven
celery
ipython
ptpython
djangorestframework-jwt

